functions/requirements.txt
azure-functions
azure-storage-blob
Pillow